feat: add rspec support
Summary:

Now you can run rspec tests with this plugin. It will work the same as all the
other runners. It will show the first line of the output in the diagnostics, I
will need to come up with some way to view the full output message.

Test Plan:

N / A, this will need to be setup and done. Right now I have no idea if this
will stick so putting time into testing will come if I plan to maintain this.