feat: add function and command to view details of the current error
Summary:

When we run our tests we are only showing the first line in the diagnostic.
This is to keep things neat as some errors can get quite long.

You now get a new command `SpecViewError` this will open a new buffer and put
the error output in it. If you have "baleia" installed it will use this to
highlight the buffer. Any ansi codes will be replaced with colors so we get a
nice view of the output.

Test Plan:

N / A, other than docs I think this will be the last thing I do before I start
setting things up to actually maintain it.