feat: add nvim busted adapter
Summary:

This adds the nvim busted adapter, this extends the busted provider to run
busted under nvim. This can be used to test nvim plugins.

There will be more docs on this another day. Right now this seems to be working
well, and I may start officially supporting this soon

Test Plan:

N / A